+++ back/src/main/java/com/tdevilleduc/urthehero/back/service/IPageService.kt
@@ -8,6 +8,7 @@ interface IPageService {
     fun exists(pageId: Int): Boolean
     fun notExists(pageId: Int): Boolean
     fun findById(pageId: Int): Page
-    fun createOrUpdate(page: PageDTO): PageDTO
+    fun createOrUpdate(pageDto: PageDTO): PageDTO
     fun delete(pageId: Int)
+    fun countByStoryId(storyId: Int): Long
 }