+++ back/src/main/java/com/tdevilleduc/urthehero/back/service/IProgressionService.kt
@@ -7,5 +7,5 @@ interface IProgressionService {
     fun doProgressionAction(personId: Int, storyId: Int, newPageId: Int): Progression
     fun findByPersonId(personId: Int): MutableList<Progression>
     fun findByPersonIdAndStoryId(personId: Int, storyId: Int): Optional<Progression>
-    fun countByStoryId(storyId: Int): Long
+    fun countPersonsByStoryId(storyId: Int): Long
 }