+++ back/src/main/java/com/tdevilleduc/urthehero/back/service/impl/PageService.kt
@@ -4,10 +4,12 @@ import com.tdevilleduc.urthehero.back.config.Mapper
 import com.tdevilleduc.urthehero.back.constant.ApplicationConstants
 import com.tdevilleduc.urthehero.back.dao.PageDao
 import com.tdevilleduc.urthehero.back.exceptions.PageNotFoundException
+import com.tdevilleduc.urthehero.back.exceptions.StoryNotFoundException
 import com.tdevilleduc.urthehero.back.model.Page
 import com.tdevilleduc.urthehero.back.model.PageDTO
 import com.tdevilleduc.urthehero.back.service.INextPageService
 import com.tdevilleduc.urthehero.back.service.IPageService
+import com.tdevilleduc.urthehero.back.service.IStoryService
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.slf4j.helpers.MessageFormatter
@@ -21,6 +23,8 @@ class PageService : IPageService {
 
     @Autowired
     private lateinit var nextPageService: INextPageService
+    @Autowired
+    private lateinit var storyService: IStoryService
     @Autowired
     private lateinit var pageDao: PageDao
 
@@ -54,8 +58,8 @@ class PageService : IPageService {
         return page
     }
 
-    override fun createOrUpdate(pageDTO: PageDTO): PageDTO {
-        val page: Page = Mapper.convert(pageDTO)
+    override fun createOrUpdate(pageDto: PageDTO): PageDTO {
+        val page: Page = Mapper.convert(pageDto)
         return Mapper.convert(pageDao.save(page))
     }
 
@@ -63,5 +67,13 @@ class PageService : IPageService {
         Assert.notNull(pageId, ApplicationConstants.CHECK_PAGEID_PARAMETER_MANDATORY!!)
         val page = findById(pageId)
         pageDao.delete(page)
+    }
+
+    override fun countByStoryId(storyId: Int): Long {
+        Assert.notNull(storyId, ApplicationConstants.CHECK_STORYID_PARAMETER_MANDATORY!!)
+        if (storyService.notExists(storyId)) {
+            throw StoryNotFoundException(MessageFormatter.format(ApplicationConstants.ERROR_MESSAGE_STORY_DOESNOT_EXIST, storyId).message)
+        }
+        return pageDao.countByStoryId(storyId)
     }
 }