+++ back/src/main/resources/data.sql
@@ -3,14 +3,14 @@ INSERT INTO person VALUES(1, 'Thomas Deville-Duc', 'thomas@gmail.com', 'tdeville
 INSERT INTO person VALUES(2, 'Marion Gianesini', 'marion@gmail.com', 'mgianesini', 'password');
 INSERT INTO person VALUES(3, 'Nicolas Danet', 'nicolas@gmail.com', 'ndanet', 'password');
 -- pageId, image, text
-INSERT INTO page(id, image, text) VALUES(1, 'image3', 'Ulysse');
-INSERT INTO page(id, image, text) VALUES(2, 'image3', 'Dès');
-INSERT INTO page(id, image, text) VALUES(3, 'image3', 'Le roman');
-INSERT INTO page(id, image, text) VALUES(4, 'image3', 'Voyage au bout de la nuit est le premier roman de Céline, publié en 1932. Ce livre manqua de deux voix le prix Goncourt mais obtient le prix Renaudot1. Il est traduit en 37 langues2.');
-INSERT INTO page(id, image, text) VALUES(5, 'image3', 'Le roman est notamment célèbre');
-INSERT INTO page(id, image, text) VALUES(6, 'image3', 'Toutefois');
-INSERT INTO page(id, image, text) VALUES(7, 'image3', 'Madame Bovary');
-INSERT INTO page(id, image, text) VALUES(8, 'image2', 'En plein centre');
+INSERT INTO page(id, story_id, image, text) VALUES(1, 1, 'image3', 'Ulysse');
+INSERT INTO page(id, story_id, image, text) VALUES(2, 1, 'image3', 'Dès');
+INSERT INTO page(id, story_id, image, text) VALUES(3, 1, 'image3', 'Le roman');
+INSERT INTO page(id, story_id, image, text) VALUES(4, 2, 'image3', 'Voyage au bout de la nuit est le premier roman de Céline, publié en 1932. Ce livre manqua de deux voix le prix Goncourt mais obtient le prix Renaudot1. Il est traduit en 37 langues2.');
+INSERT INTO page(id, story_id, image, text) VALUES(5, 2, 'image3', 'Le roman est notamment célèbre');
+INSERT INTO page(id, story_id, image, text) VALUES(6, 2, 'image3', 'Toutefois');
+INSERT INTO page(id, story_id, image, text) VALUES(7, 3, 'image3', 'Madame Bovary');
+INSERT INTO page(id, story_id, image, text) VALUES(8, 1, 'image2', 'En plein centre');
 -- storyId, authorId, firstPageId, title
 INSERT INTO story(story_Id, author_Id, first_Page_Id, title, detailed_Text, image) VALUES(1, 1, 1, 'Ulysse', 'blablabla Ulysse prenons un texte long pour décrire lhistoire', 'imageUlysse');
 INSERT INTO story(story_Id, author_Id, first_Page_Id, title, detailed_Text, image) VALUES(2, 2, 4, 'Voyage au bout de la nuit', 'bliblibli voyage voyage !!', 'imageVoyage');