+++ back/src/test/java/com/tdevilleduc/urthehero/back/model/PageTest.kt
@@ -10,11 +10,14 @@ internal class PageTest {
         val pageId = 46
         val pageText = "test de titre"
         val pageImage = "image de ouf"
+        val pageStoryId = 27
         val page = Page(pageText, pageImage)
         page.id = pageId
+        page.storyId = pageStoryId
         Assertions.assertEquals(page.id!!, pageId)
         Assertions.assertEquals(page.text, pageText)
         Assertions.assertEquals(page.image, pageImage)
+        Assertions.assertEquals(page.storyId!!, pageStoryId)
         Assertions.assertTrue(page.nextPageList.isEmpty())
     }
 }